estaba una variable como int y tenia que ser Long

+++ cabana/src/main/java/com/proyecto/entities/client/CountClient.java
@@ -4,19 +4,19 @@ import com.proyecto.entities.Client;
 
 public class CountClient {
 
-    private Integer total;
+    private Long total;
     private Client client;
 
-    public CountClient(Integer total, Client client) {
+    public CountClient(Long total, Client client) {
         this.total = total;
         this.client = client;
     }
 
-    public Integer getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
